FST-115 move shared utilities
