+++ src/features/integration/components/form/ApplicantForm.tsx
@@ -1,6 +1,6 @@
 import {FormGroup} from '@mui/material';
 import React, {useContext} from 'react';
-import {applicantOptions} from "../../util/DefaultValues";
+import {applicantOptions} from "../../defaults/DefaultValues";
 import InputField from "./InputField";
 import {INPUT_TYPE} from "../../types/InputType.enum";
 import {IInputField} from "../../types/InputField";
